% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_data}
\alias{get_data}
\title{query the at-sea Access database for catch data}
\usage{
get_data(channel = channel, sta_num = 1)
}
\arguments{
\item{channel}{channel to the at-sea database}

\item{sta_num}{station number}
}
\description{
query the at-sea Access database for catch data
}
\examples{
\dontrun{
get_data(channel = channel, sta_num = 1)
}
}
