% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{make_plot_data}
\alias{make_plot_data}
\title{prepare data for plots}
\usage{
make_plot_data(depth, cat, strat, names)
}
\arguments{
\item{depth}{depth by skate}

\item{cat}{catch by skate}

\item{strat}{depth strata}

\item{names}{names of the species from codes}
}
\value{
list of data used for making plots
}
\description{
prepare data for plots
}
\examples{
\dontrun{
make_plot_data(depth, cat, strat, names)
}
}
