% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_data}
\alias{plot_data}
\title{make and save plots for use in assessing whale depredation}
\usage{
plot_data(
  depth_strat = depth_strat,
  spp_sum_strat = spp_sum_strat,
  dep = dep,
  catch = catch
)
}
\arguments{
\item{depth_strat}{depth strata distinctions}

\item{spp_sum_strat}{summary of catch by species and depth strata}

\item{dep}{depredation evidence}

\item{catch}{catch data by skate}
}
\description{
make and save plots for use in assessing whale depredation
}
\examples{
\dontrun{
plot_data(depth_strat = depth_strat, spp_sum_strat = spp_sum_strat, dep = dep, catch = catch
}
}
