% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whale_dep.R
\name{whale_dep}
\alias{whale_dep}
\title{wrapper function to produce plots}
\usage{
whale_dep(channel, station)
}
\arguments{
\item{channel}{where the at-sea Access database is}

\item{station}{station number}
}
\value{
plots for viewing
}
\description{
wrapper function to produce plots
}
\examples{
\dontrun{
whale_dep(channel, station)
}
}
